requirements.txt
alembic==0.9.1
appdirs==1.4.3
APScheduler==3.3.1
click==6.7
Flask==0.12
Flask-Admin==1.5.0
Flask-APScheduler==1.7.0
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==1.0
mysql-connector-python-rf==2.2.2
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2016.10
requests==2.13.0
six==1.10.0
slackclient==1.0.5
SQLAlchemy==1.1.6
tzlocal==1.3
websocket-client==0.40.0
Werkzeug==0.12.1
WTForms==2.1
